refactor [mrmustard] - create the `training` package
