+++ mrmustard/lab/states.py
@@ -16,12 +16,12 @@
 This module implements the quantum states upon which a quantum circuits acts on.
 """
 
-from typing import Union, Optional, List, Tuple, Sequence, Dict
-from mrmustard.types import Scalar, Vector, Matrix, Trainable
+from typing import Union, Optional, List, Tuple, Sequence
+from mrmustard.types import Scalar, Vector, Matrix
 from mrmustard import settings
 from mrmustard.lab.abstract import State
 from mrmustard.physics import gaussian, fock
-from mrmustard.utils.parametrized import Parametrized
+from mrmustard.training import Parametrized
 from mrmustard.math import Math
 
 math = Math()
