+++ mrmustard/training/parametrized.py
@@ -18,7 +18,7 @@ an abstract base class for all parametrized objects.
 
 from mrmustard.types import Sequence, List, Generator, Iterable, Any
 from mrmustard.math import Math
-from mrmustard.utils.parameter import create_parameter, Trainable, Constant, Parameter
+from .parameter import create_parameter, Trainable, Constant, Parameter
 
 math = Math()
